the `bot/login`
