+++ bot/login/auth.go
@@ -1,0 +1,98 @@
+package login
+
+import (
+	"bytes"
+	"crypto/sha1"
+	"encoding/json"
+	"fmt"
+	"github.com/Tnze/go-mc/bot"
+	"io"
+	"net/http"
+	"strings"
+)
+
+// AuthDigest computes a special SHA-1 digest required for Minecraft web
+// authentication on Premium servers (online-mode=true).
+// Source: http://wiki.vg/Protocol_Encryption#Server
+//
+// Also many, many thanks to SirCmpwn and his wonderful gist (C#):
+// https://gist.github.com/SirCmpwn/404223052379e82f91e6
+func AuthDigest(serverID string, sharedSecret, publicKey []byte) string {
+	h := sha1.New()
+	h.Write([]byte(serverID))
+	h.Write(sharedSecret)
+	h.Write(publicKey)
+	hash := h.Sum(nil)
+
+	// Check for negative hashes
+	negative := (hash[0] & 0x80) == 0x80
+	if negative {
+		hash = twosComplement(hash)
+	}
+
+	// Trim away zeroes
+	res := strings.TrimLeft(fmt.Sprintf("%x", hash), "0")
+	if negative {
+		res = "-" + res
+	}
+
+	return res
+}
+
+// little endian
+func twosComplement(p []byte) []byte {
+	carry := true
+	for i := len(p) - 1; i >= 0; i-- {
+		p[i] = ^p[i]
+		if carry {
+			carry = p[i] == 0xff
+			p[i]++
+		}
+	}
+	return p
+}
+
+type request struct {
+	AccessToken     string `json:"accessToken"`
+	SelectedProfile string `json:"selectedProfile"`
+	ServerID        string `json:"serverId"`
+}
+
+func LoginAuth(auth bot.Auth, shareSecret []byte, er EncryptionRequest) error {
+	digest := AuthDigest(er.ServerID, shareSecret, er.PublicKey)
+	return joinServer(auth.AsTk, auth.UUID, digest)
+}
+
+func joinServer(asTk, uuid, serverId string) error {
+	requestContent, err := json.Marshal(
+		request{
+			AccessToken:     asTk,
+			SelectedProfile: uuid,
+			ServerID:        serverId,
+		},
+	)
+	if err != nil {
+		return fmt.Errorf("create request packet to yggdrasil faile: %v", err)
+	}
+
+	client := http.Client{}
+	PostRequest, err := http.NewRequest(http.MethodPost, "https://sessionserver.mojang.com/session/minecraft/join",
+		bytes.NewReader(requestContent))
+	if err != nil {
+		return fmt.Errorf("make request error: %v", err)
+	}
+	PostRequest.Header.Set("User-agent", "go-mc")
+	PostRequest.Header.Set("Connection", "keep-alive")
+	PostRequest.Header.Set("Content-Type", "application/json")
+	resp, err := client.Do(PostRequest)
+	if err != nil {
+		return fmt.Errorf("post fail: %v", err)
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusNoContent {
+		body, _ := io.ReadAll(resp.Body)
+		return fmt.Errorf("auth fail: %s", string(body))
+	}
+	return nil
+}
